+++ adapters/lumi/WXCJKG12LM.py
@@ -2,7 +2,7 @@
 from adapters.lumi.aqara_opple_switch import AqaraOppleSwitch
 
 
-class WXCJKG13LM(AqaraOppleSwitch):
+class WXCJKG12LM(AqaraOppleSwitch):
     def __init__(self, devices):
-        buttons_count = 6
+        buttons_count = 4
         super().__init__(devices, buttons_count)